decaf377::fields::fq
Type Alias Fq

pub type Fq = Fq;

Aliased Type

struct Fq(/* private fields */);

Implementations

impl Fq

pub const MODULUS_LIMBS: [u64; 4] = _

pub const MODULUS_MINUS_ONE_DIV_TWO_LIMBS: [u64; 4] = _

pub const MODULUS_BIT_SIZE: u32 = 253u32

pub const TRACE_LIMBS: [u64; 4] = _

pub const TRACE_MINUS_ONE_DIV_TWO_LIMBS: [u64; 4] = _

pub const TWO_ADICITY: u32 = 47u32

pub const QUADRATIC_NON_RESIDUE_TO_TRACE: Self = _

pub const MULTIPLICATIVE_GENERATOR: Self = _

pub const TWO_ADIC_ROOT_OF_UNITY: Self = _

pub const FIELD_SIZE_POWER_OF_TWO: Self = _

pub fn from_le_bytes_mod_order(bytes: &[u8]) -> Self

pub fn from_bytes_checked(bytes: &[u8; 32]) -> Result<Self, EncodingError>

Convert bytes into an Fq element, returning None if these bytes are not already reduced.

This means that values that cannot be produced by encoding a field element will return None, enforcing canonical serialization.

pub fn to_bytes(&self) -> [u8; 32]

pub fn rand<R: CryptoRngCore>(rng: &mut R) -> Self

Sample a random field element uniformly.

pub fn power<S: AsRef<[u64]>>(&self, exp: S) -> Self

Raise this element to a given power.

Note: Arkworks provides another method for this, called pow.

impl Fq

pub fn sqrt_ratio_zeta(num: &Self, den: &Self) -> (bool, Self)

Computes the square root of a ratio of field elements, returning:

(true, sqrt(num/den)) if num and den are both nonzero and num/den is square;

(true, 0) if num is zero;

(false, 0) if den is zero;

(false, sqrt(zeta*num/den)) if num and den are both nonzero and num/den is nonsquare;