decaf377::fields::fr::u64::wrapper
Struct Fr

pub struct Fr(/* private fields */);

Implementations

impl Fr

pub const ZERO: Self = _

pub const ONE: Self = _

pub fn to_bytes_le(&self) -> [u8; 32]

pub fn square(&self) -> Fr

pub fn inverse(&self) -> Option<Fr>

pub fn add(self, other: &Fr) -> Fr

pub fn sub(self, other: &Fr) -> Fr

pub fn mul(self, other: &Fr) -> Fr

pub fn neg(self) -> Fr

impl Fr

pub const MODULUS_LIMBS: [u64; 4] = _

pub const MODULUS_MINUS_ONE_DIV_TWO_LIMBS: [u64; 4] = _

pub const MODULUS_BIT_SIZE: u32 = 251u32

pub const TRACE_LIMBS: [u64; 4] = _

pub const TRACE_MINUS_ONE_DIV_TWO_LIMBS: [u64; 4] = _

pub const TWO_ADICITY: u32 = 1u32

pub const MULTIPLICATIVE_GENERATOR: Self = _

pub const TWO_ADIC_ROOT_OF_UNITY: Self = _

pub const FIELD_SIZE_POWER_OF_TWO: Self = _

pub fn from_le_bytes_mod_order(bytes: &[u8]) -> Self

pub fn from_bytes_checked(bytes: &[u8; 32]) -> Result<Self, EncodingError>

Convert bytes into an Fr element, returning None if these bytes are not already reduced.

This means that values that cannot be produced by encoding a field element will return None, enforcing canonical serialization.

pub fn to_bytes(&self) -> [u8; 32]

pub fn rand<R: CryptoRngCore>(rng: &mut R) -> Self

Sample a random field element uniformly.