decaf377::r1cs
Type Alias FqVar

pub type FqVar = FpVar<Fq>;

Aliased Type

enum FqVar {
    Constant(Fq),
    Var(AllocatedFp<Fq>),
}

Variants

Constant(Fq)

Represents a constant in the constraint system, which means that it does not have a corresponding variable.

Var(AllocatedFp<Fq>)

Represents an allocated variable constant in the constraint system.