decaf377
Struct Element

pub struct Element { /* private fields */ }

Implementations

impl Element

pub const GENERATOR: Self = _

Return the conventional generator for decaf377.

pub const IDENTITY: Self = _

impl Element

pub fn is_identity(&self) -> bool

Convenience method to make identity checks more readable.

pub fn vartime_multiscalar_mul<I, J>(scalars: I, points: J) -> Element
where I: IntoIterator, I::Item: Borrow<Fr>, J: IntoIterator, J::Item: Borrow<Element>,

Given an iterator of public scalars and an iterator of public points, compute $$ Q = [c_1] P_1 + \cdots + [c_n] P_n, $$ using variable-time operations.

It is an error to call this function with two iterators of different lengths – it would require ExactSizeIterator, but ExactSizeIterators are not closed under chaining, and disallowing iterator chaining would destroy the utility of the function.

impl Element

pub fn hash_to_curve(r_1: &Fq, r_2: &Fq) -> Element

Maps two field elements to a uniformly distributed decaf377 Element.

The two field elements provided as inputs should be independently chosen.

pub fn encode_to_curve(r: &Fq) -> Element

Maps a field element to a decaf377 Element suitable for CDH challenges.

impl Element

pub fn negate(&self) -> Element

pub fn vartime_compress_to_field(&self) -> Fq

pub fn vartime_compress(&self) -> Encoding